backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.1.3
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.2
python-multipart==0.0.6
aiofiles==23.2.1
python-magic-bin==0.4.14
opensearch-py==2.3.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.1
jinja2==3.1.2
email-validator==2.1.0.post1
PyPDF2==3.0.1
requests==2.31.0